turn/config: cleanup
